+++ wormhole-connect/src/utils/wallet/ReadOnlyWallet.ts
@@ -8,16 +8,12 @@ import {
 import { Chain, chainToChainId } from '@wormhole-foundation/sdk';
 
 export class ReadOnlyWallet extends Wallet {
-  private readonly _address: string;
-  private readonly _chain: Chain;
   private _isConnected: boolean = true;
 
   static readonly NAME = 'ReadyOnlyWallet';
 
-  constructor(address: string, chain: Chain) {
+  constructor(readonly _address: string, readonly _chain: Chain) {
     super();
-    this._address = address;
-    this._chain = chain;
   }
 
   getName(): string {
